SHM5: Class Coordinates - positions changed from size_t to int

+++ shm/coordinates.cpp
@@ -1,6 +1,6 @@
 #include "coordinates.hpp"
 
-Coordinates::Coordinates(size_t positionX, size_t positionY)
+Coordinates::Coordinates(int positionX, int positionY)
     : positionX_(positionX), positionY_(positionY) {}
 
 bool Coordinates::operator==(const Coordinates& rhs) const {
